Extract request step execution
